% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FingerASA.R
\name{UCM}
\alias{UCM}
\title{Usage: load(subjects=c("Paula","Angelo"),wd=c("/Users/angelobj/Box Sync/ASA old paper/Prueba/"),Task="ASA")}
\usage{
UCM(subjects, wd = NULL, Task = NULL, timeR = c(5.5, 7.5), timeS = c(6,
  20), freq = 1000, cutfreq = 0.02, order = 2, J = matrix(c(1, 1, 1, 1),
  ncol = 1))
}
\description{
Usage: load(subjects=c("Paula","Angelo"),wd=c("/Users/angelobj/Box Sync/ASA old paper/Prueba/"),Task="ASA")
}
